VisualizationsController#create powered by ActiveRecord models
