+++ app/controllers/api/json/visualizations_controller.rb
@@ -4,8 +4,6 @@ require_relative '../../../models/visualization/member'
 require_relative '../../../models/visualization/collection'
 require_relative '../../../models/visualization/presenter'
 require_relative '../../../models/visualization/locator'
-require_relative '../../../models/visualization/name_generator'
-require_relative '../../../models/visualization/table_blender'
 require_relative '../../../models/visualization/watcher'
 require_relative '../../../models/map/presenter'
 require_relative '../../../../lib/static_maps_url_helper'
@@ -26,76 +24,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
   before_filter :table_and_schema_from_params, only: [:update, :stats,
                                                       :notify_watching, :list_watching,
                                                       :add_like, :remove_like, :set_next_id]
-
-  def create
-    vis_data = payload
-
-    vis_data.delete(:permission)
-    vis_data.delete(:permission_id)
-
-    # Don't allow to modify next_id/prev_id, force to use set_next_id()
-    prev_id = vis_data.delete(:prev_id) || vis_data.delete('prev_id')
-    next_id = vis_data.delete(:next_id) || vis_data.delete('next_id')
-
-    param_tables = params[:tables]
-    current_user_id = current_user.id
-
-    origin = 'blank'
-    source_id = params[:source_visualization_id]
-    vis = if source_id
-            user = Carto::User.find(current_user_id)
-            source = Carto::Visualization.where(id: source_id).first
-            return head(403) unless source && source.is_viewable_by_user?(user) && !source.kind_raster?
-            if source.derived?
-              origin = 'copy'
-              duplicate_derived_visualization(params[:source_visualization_id], user)
-            else
-              create_visualization_from_tables([source.user_table], vis_data)
-            end
-          elsif param_tables
-            subdomain = CartoDB.extract_subdomain(request)
-            viewed_user = ::User.find(username: subdomain)
-            tables = param_tables.map do |table_name|
-              Carto::Helpers::TableLocator.new.get_by_id_or_name(table_name, viewed_user) if viewed_user
-            end
-            create_visualization_from_tables(tables.flatten, vis_data)
-          else
-            Visualization::Member.new(vis_data.merge(name: name_candidate, user_id: current_user_id))
-          end
-
-    vis.ensure_valid_privacy
-    # both null, make sure is the first children or automatically link to the tail of the list
-    if !vis.parent_id.nil? && prev_id.nil? && next_id.nil?
-      parent_vis = Visualization::Member.new(id: vis.parent_id).fetch
-      return head(403) unless parent_vis.has_permission?(current_user, Visualization::Member::PERMISSION_READWRITE)
-
-      children = parent_vis.children
-
-      prev_id = children.last.id unless children.empty?
-    end
-
-    vis.store
-
-    vis = set_visualization_prev_next(vis, prev_id, next_id)
-
-    current_viewer_id = current_viewer.id
-    properties = {
-      user_id: current_viewer_id,
-      origin: origin,
-      visualization_id: vis.id
-    }
-
-    if vis.derived?
-      Carto::Tracking::Events::CreatedMap.new(current_viewer_id, properties).report
-    else
-      Carto::Tracking::Events::CreatedDataset.new(current_viewer_id, properties).report
-    end
-
-    render_jsonp(vis)
-  rescue CartoDB::InvalidMember => e
-    CartoDB::Logger.error(message: "Invalid member creating visualization", visualization_id: vis.id, exception: e)
-    render_jsonp({ errors: vis.full_errors }, 400)
-  end
 
   def update
     @stats_aggregator.timing('visualizations.update') do
@@ -280,37 +208,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
 
   private
 
-  def duplicate_derived_visualization(source, user)
-    visualization_copy_id = @stats_aggregator.timing('copy') do
-      export_service = Carto::VisualizationsExportService2.new
-      visualization_hash = export_service.export_visualization_json_hash(source, user)
-      visualization_copy = export_service.build_visualization_from_hash_export(visualization_hash)
-      visualization_copy.name = name_candidate
-      visualization_copy.version = user.new_visualizations_version
-      Carto::VisualizationsExportPersistenceService.new.save_import(user, visualization_copy)
-      visualization_copy.id
-    end
-
-    CartoDB::Visualization::Member.new(id: visualization_copy_id).fetch
-  end
-
-  def create_visualization_from_tables(tables, vis_data)
-    blender = Visualization::TableBlender.new(Carto::User.find(current_user.id), tables)
-    map = blender.blend
-
-    vis = Visualization::Member.new(vis_data.merge(name: name_candidate,
-                                                   map_id: map.id,
-                                                   type: 'derived',
-                                                   privacy: blender.blended_privacy,
-                                                   user_id: current_user.id))
-
-    @stats_aggregator.timing('default-overlays') do
-      Visualization::Overlays.new(vis).create_default_overlays
-    end
-
-    vis
-  end
-
   def table_and_schema_from_params
     if params.fetch('id', nil) =~ /\./
       @table_id, @schema = params.fetch('id').split('.').reverse
@@ -350,11 +247,6 @@ class Api::Json::VisualizationsController < Api::ApplicationController
   def payload
     request.body.rewind
     ::JSON.parse(request.body.read.to_s || String.new, {symbolize_names: true})
-  end
-
-  def name_candidate
-    Visualization::NameGenerator.new(current_user)
-                                .name(params[:name])
   end
 
   def tables_by_map_id(map_ids)
@@ -397,18 +289,5 @@ class Api::Json::VisualizationsController < Api::ApplicationController
     elsif vis.type == Carto::Visualization::TYPE_DERIVED
       ::Resque.enqueue(::Resque::UserJobs::Mail::MapLiked, vis.id, current_viewer.id, vis_preview_image)
     end
-  end
-
-  def set_visualization_prev_next(vis, prev_id, next_id)
-    if !prev_id.nil?
-      prev_vis = Visualization::Member.new(id: prev_id).fetch
-      return head(403) unless prev_vis.has_permission?(current_user, Visualization::Member::PERMISSION_READWRITE)
-      prev_vis.set_next_list_item!(vis)
-    elsif !next_id.nil?
-      next_vis = Visualization::Member.new(id: next_id).fetch
-      return head(403) unless next_vis.has_permission?(current_user, Visualization::Member::PERMISSION_READWRITE)
-      next_vis.set_prev_list_item!(vis)
-    end
-    vis
   end
 end
